Rename JsonHomeService to *Client, add AsyncJsonHomeClient.
The JsonHomeClient builder with a WSClient only now
builds the AsyncJsonHomeClient.

The former JsonHomeClient functionality was just inlined
into the new JsonHomeClient (former JsonHomeService).

Cleanup needed: The Builder.addHost() accepts host:String
and LinkRelations*, while JsonHomeClient.getUrl expects
a JsonHomeHost and the LinkRelation.
One of both should be changed.

The README also still needs some love, it does not yet
reflect the changes.
