+++ src/main/scala/de/kaufhof/jsonhomeclient/JsonHomeClient.scala
@@ -16,33 +16,234 @@
  */
 package de.kaufhof.jsonhomeclient
 
-import scala.language.postfixOps
+import akka.actor.ActorSystem
+import com.damnhandy.uri.template.UriTemplate
+import play.api.libs.ws.WSClient
 
-import play.api.libs.json.JsValue
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext.Implicits.global
-import play.api.libs.ws._
+import scala.concurrent.duration.{FiniteDuration, Duration}
+import scala.language.implicitConversions
 
 /**
- * The client to load a json-home document and extract relevant information.
- * A client is responsible for a specific json-home host.
- *
- * @param host the json home host to load the home document from
- * @param ws the WSClient to use for loading the home document.
- *           In a Play 2.3.x application it can be created via `WS.client`,
- *           in other apps it can be created via
- *           `new NingWSClient(new AsyncHttpClientConfig.Builder().build())` (remember
- *           to `close()` the client when the application is stopped).
+ * The json-home service allows to resolve urls (href/href-template) for a given json-home host and a given
+ * link releation type.
  *
  * @author <a href="mailto:martin.grotzke@inoio.de">Martin Grotzke</a>
  */
-class JsonHomeClient(val host: JsonHomeHost,
-                     val ws: WSClient) {
+case class JsonHomeClient(cachesByHost: Map[JsonHomeHost, JsonHomeCache]) {
 
-  private[jsonhomeclient] def jsonHome(): Future[JsValue] = {
-    ws.url(host.jsonHomeUri.toString)
-      .withHeaders("Accept" -> "application/json-home").get().map(_.json)
+  /**
+   * Determines the url (json-home "href") for the given json home host and the given direct link relation.
+   */
+  def getUrl(host: JsonHomeHost, relation: DirectLinkRelationType): Option[String] = {
+    cachesByHost.get(host).flatMap(_.getUrl(relation))
+  }
+
+  /**
+   * Determines the url (json-home "href-template") for the given json home host and the given template link relation.
+   * The href template variables are replaced using the provided params.
+   */
+  def getUrl(host: JsonHomeHost, relation: TemplateLinkRelationType, params: Map[String, Any]): Option[String] = {
+    cachesByHost.get(host).flatMap(_.getUrl(relation)).map { hrefTemplate =>
+      params.foldLeft(UriTemplate.fromTemplate(hrefTemplate))((res, param) => res.set(param._1, param._2)).expand()
+    }
   }
 
 }
 
+object JsonHomeClient {
+
+  sealed trait SystemFlag
+  trait WithSystem extends SystemFlag
+  trait WithoutSystem extends SystemFlag
+
+  sealed trait HostFlag
+  trait WithHost extends HostFlag
+  trait WithoutHost extends HostFlag
+
+
+
+
+  sealed trait WSClientFlag
+  trait WithWSClient extends WSClientFlag
+  trait WithoutWSClient extends WSClientFlag
+
+  sealed trait UpdateIntervalFlag
+  trait WithUpdateInterval extends UpdateIntervalFlag
+  trait WithoutUpdateInterval extends UpdateIntervalFlag
+
+  sealed trait StartDelayFlag
+  trait WithStartDelay extends StartDelayFlag
+  trait WithoutStartDelay extends StartDelayFlag
+
+
+
+  object JsonHomeClientBuilder{
+    implicit def enableCachedClientBuild[A <: WithWSClient, B <: WithSystem, C <: WithUpdateInterval, D <: WithStartDelay, H >: HostsExist[A, B, C, D]](builder: Builder[A, B, C, D] with H): {def build():JsonHomeClient} = new {
+      def build(): JsonHomeClient = {
+        val caches = for{
+          host <- builder.hosts
+          client <- builder.wsClient
+          system <- builder.system
+          interval <- builder.updateInterval
+          delay <- builder.startDelay
+        } yield {
+          new JsonHomeCache(host, client, system, interval, delay)
+        }
+
+        JsonHomeClient(caches)
+      }
+    }
+
+    implicit def enableBuildWithDefaultStartDelay[A <: WithWSClient, B <: WithSystem, C <: WithUpdateInterval, D <: WithoutStartDelay, H >: HostsExist[A, B, C, D]](builder: Builder[A, B, C, D] with H): {def build():JsonHomeClient} = new {
+      def build(): JsonHomeClient = {
+        val caches = for{
+          host <- builder.hosts
+          client <- builder.wsClient
+          system <- builder.system
+          interval <- builder.updateInterval
+        } yield {
+          new JsonHomeCache(host, client, system, updateInterval = interval)
+        }
+
+        JsonHomeClient(caches)
+      }
+    }
+
+    implicit def enableBuildWithDefaultUpdateInterval[A <: WithWSClient, B <: WithSystem, C <: WithoutUpdateInterval, D <: WithStartDelay, H >: HostsExist[A, B, C, D]](builder: Builder[A, B, C, D] with H): {def build():JsonHomeClient} = new {
+      def build(): JsonHomeClient = {
+        val caches = for{
+          host <- builder.hosts
+          client <- builder.wsClient
+          system <- builder.system
+          delay <- builder.startDelay
+        } yield {
+          new JsonHomeCache(host, client, system, initialTimeToWait = delay)
+        }
+
+        JsonHomeClient(caches)
+      }
+    }
+
+    implicit def enableBuildWithDefaultDurations[A <: WithWSClient, B <: WithSystem, C <: WithoutUpdateInterval, D <: WithoutStartDelay, H >: HostsExist[A, B, C, D]](builder: Builder[A, B, C, D] with H): {def build():JsonHomeClient} = new {
+      def build(): JsonHomeClient = {
+        val caches = for{
+          host <- builder.hosts
+          client <- builder.wsClient
+          system <- builder.system
+        } yield {
+          new JsonHomeCache(host, client, system)
+        }
+
+        JsonHomeClient(caches)
+      }
+    }
+
+    implicit def enableAsyncClientBuild[A <: WithWSClient, B <: WithoutSystem, C <: WithoutUpdateInterval, D <: WithoutStartDelay](builder: Builder[A, B, C, D]): {def build():AsyncJsonHomeClient} = new {
+      def build(): AsyncJsonHomeClient = {
+        new AsyncJsonHomeClient(builder.wsClient.get)
+      }
+    }
+  }
+
+  trait Builder[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag]{
+    def hosts: List[JsonHomeHost]
+    def wsClient: Option[WSClient]
+    def system: Option[ActorSystem]
+    def updateInterval: Option[FiniteDuration]
+    def startDelay: Option[FiniteDuration]
+
+    def withWSClient(client: WSClient): Builder[WithWSClient, SF, UF, SDF] with HostHandling[WithWSClient, SF, UF, SDF]
+    def withStartDelay(delay: FiniteDuration): Builder[CF, SF, UF, WithStartDelay] with HostHandling[CF, SF, UF, WithStartDelay]
+    def withUpdateInterval(updateInterval: FiniteDuration): Builder[CF, SF, WithUpdateInterval, SDF] with HostHandling[CF, SF, WithUpdateInterval, SDF]
+    def withCaching(system: ActorSystem): Builder[CF, WithSystem, UF, SDF] with HostHandling[CF, WithSystem, UF, SDF]
+  }
+
+  trait HostHandling[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag]{self: Builder[CF, SF, UF, SDF] =>}
+
+  trait ZeroHosts[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] extends HostHandling[CF, SF, UF, SDF]{self: Builder[CF, SF, UF, SDF] =>
+    def addHost(url: String, rels: LinkRelationType*): JsonHomeServiceBuilderOne[CF, SF, UF, SDF] with OneHost[CF, SF, UF, SDF] = {
+      val host = JsonHomeHost(url, rels)
+      JsonHomeServiceBuilderOne(host :: hosts, wsClient, system, updateInterval, startDelay)
+    }
+  }
+
+  trait HostsExist[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] extends HostHandling[CF, SF, UF, SDF]{self: Builder[CF, SF, UF, SDF] =>
+    def addHost(url: String, rels: LinkRelationType*): JsonHomeServiceBuilderMany[CF, SF, UF, SDF] with ManyHosts[CF, SF, UF, SDF] = {
+      val host = JsonHomeHost(url, rels)
+      JsonHomeServiceBuilderMany(host :: hosts, wsClient, system, updateInterval, startDelay)
+    }
+  }
+
+  trait OneHost[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] extends HostsExist[CF, SF, UF, SDF]{self: Builder[CF, SF, UF, SDF] =>}
+
+  trait ManyHosts[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] extends HostsExist[CF, SF, UF, SDF]{self: Builder[CF, SF, UF, SDF] =>}
+
+
+
+  case class JsonHomeServiceBuilderZero[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] private[jsonhomeclient](
+                                     hosts: List[JsonHomeHost] = Nil,
+                                     wsClient: Option[WSClient] = None,
+                                     system: Option[ActorSystem] = None,
+                                     updateInterval: Option[FiniteDuration] = None,
+                                     startDelay: Option[FiniteDuration] = None) extends Builder[CF, SF, UF, SDF] with ZeroHosts[CF, SF, UF, SDF]{
+
+
+    def withWSClient(client: WSClient) = this.copy[WithWSClient, SF, UF, SDF](wsClient = Some(client))
+
+    def withStartDelay(delay: FiniteDuration) = this.copy[CF, SF, UF, WithStartDelay](startDelay = Some(delay))
+
+    def withCaching(system: ActorSystem) = this.copy[CF, WithSystem, UF, SDF](system = Some(system))
+
+    def withUpdateInterval(updateInterval: FiniteDuration) = this.copy[CF, SF, WithUpdateInterval, SDF](updateInterval = Some(updateInterval))
+
+  }
+
+  case class JsonHomeServiceBuilderOne[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] private[jsonhomeclient](
+                                                                                      hosts: List[JsonHomeHost] = Nil,
+                                                                                      wsClient: Option[WSClient] = None,
+                                                                                      system: Option[ActorSystem] = None,
+                                                                                      updateInterval: Option[FiniteDuration] = None,
+                                                                                      startDelay: Option[FiniteDuration] = None) extends Builder[CF, SF, UF, SDF] with OneHost[CF, SF, UF, SDF]{
+
+
+    def withWSClient(client: WSClient): JsonHomeServiceBuilderOne[WithWSClient, SF, UF, SDF] with OneHost[WithWSClient, SF, UF, SDF] = this.copy[WithWSClient, SF, UF, SDF](wsClient = Some(client))
+
+    def withStartDelay(delay: FiniteDuration): JsonHomeServiceBuilderOne[CF, SF, UF, WithStartDelay] with OneHost[CF, SF, UF, WithStartDelay] = this.copy[CF, SF, UF, WithStartDelay](startDelay = Some(delay))
+
+    def withCaching(system: ActorSystem): JsonHomeServiceBuilderOne[CF, WithSystem, UF, SDF] with OneHost[CF, WithSystem, UF, SDF] = this.copy[CF, WithSystem, UF, SDF](system = Some(system))
+
+    def withUpdateInterval(updateInterval: FiniteDuration): JsonHomeServiceBuilderOne[CF, SF, WithUpdateInterval, SDF] with OneHost[CF, SF, WithUpdateInterval, SDF] = this.copy[CF, SF, WithUpdateInterval, SDF](updateInterval = Some(updateInterval))
+
+  }
+
+  case class JsonHomeServiceBuilderMany[CF <: WSClientFlag, SF <: SystemFlag, UF <: UpdateIntervalFlag, SDF <: StartDelayFlag] private[jsonhomeclient](
+                                                                                     hosts: List[JsonHomeHost] = Nil,
+                                                                                     wsClient: Option[WSClient] = None,
+                                                                                     system: Option[ActorSystem] = None,
+                                                                                     updateInterval: Option[FiniteDuration] = None,
+                                                                                     startDelay: Option[FiniteDuration] = None) extends Builder[CF, SF, UF, SDF] with ManyHosts[CF, SF, UF, SDF]{
+
+
+    def withWSClient(client: WSClient): JsonHomeServiceBuilderMany[WithWSClient, SF, UF, SDF] with ManyHosts[WithWSClient, SF, UF, SDF]  = this.copy[WithWSClient, SF, UF, SDF](wsClient = Some(client))
+
+    def withStartDelay(delay: FiniteDuration): JsonHomeServiceBuilderMany[CF, SF, UF, WithStartDelay] with ManyHosts[CF, SF, UF, WithStartDelay] = this.copy[CF, SF, UF, WithStartDelay](startDelay = Some(delay))
+
+    def withCaching(system: ActorSystem): JsonHomeServiceBuilderMany[CF, WithSystem, UF, SDF] with ManyHosts[CF, WithSystem, UF, SDF]  = this.copy[CF, WithSystem, UF, SDF](system = Some(system))
+
+    def withUpdateInterval(updateInterval: FiniteDuration): JsonHomeServiceBuilderMany[CF, SF, WithUpdateInterval, SDF] with ManyHosts[CF, SF, WithUpdateInterval, SDF]  = this.copy[CF, SF, WithUpdateInterval, SDF](updateInterval = Some(updateInterval))
+
+  }
+
+
+  object Builder {
+    def apply(): Builder[WithoutWSClient, WithoutSystem, WithoutUpdateInterval, WithoutStartDelay] with ZeroHosts[WithoutWSClient, WithoutSystem, WithoutUpdateInterval, WithoutStartDelay] = JsonHomeServiceBuilderZero()
+  }
+
+  def apply(caches: Seq[JsonHomeCache]): JsonHomeClient = {
+    val cachesByHost = caches.foldLeft(Map.empty[JsonHomeHost, JsonHomeCache]) { (res, cache) =>
+      res + (cache.host -> cache)
+    }
+    new JsonHomeClient(cachesByHost)
+  }
+
+}
