+++ src/test/scala/de/kaufhof/jsonhomeclient/AsyncJsonHomeClientSpec.scala
@@ -1,18 +1,19 @@
 package de.kaufhof.jsonhomeclient
 
+import java.net.InetSocketAddress
+
+import akka.actor.ActorSystem
 import com.ning.http.client.AsyncHttpClientConfig
-import play.api.libs.ws.ning.NingWSClient
-
-import scala.language.postfixOps
-
+import com.sun.net.httpserver.{HttpExchange, HttpHandler, HttpServer}
 import org.scalatest.ConfigMap
 import play.api.libs.json.Json
-import com.sun.net.httpserver.{HttpServer, HttpExchange, HttpHandler}
-import java.net.InetSocketAddress
-import scala.concurrent.duration._
-import akka.actor.ActorSystem
+import play.api.libs.ws.ning.NingWSClient
+import play.api.test.{DefaultAwaitTimeout, FutureAwaits}
 
-class JsonHomeCacheSpec extends IntegrationSpec {
+import scala.concurrent.duration._
+import scala.language.postfixOps
+
+class AsyncJsonHomeClientSpec extends IntegrationSpec with FutureAwaits with DefaultAwaitTimeout {
 
   private var json = Json.parse(
     """
@@ -22,9 +23,9 @@ class JsonHomeCacheSpec extends IntegrationSpec {
       |	     "href": "/artists"
       |	   },
       |    "http://spec.example.org/rels/artist": {
-      |	     "href-template": "/artists/{artist-id}",
+      |	     "href-template": "/artists/{artist_id}",
       |      "href-vars": {
-      |        "artist-id": "http://spec.example.org/params/artist"
+      |        "artist_id": "http://spec.example.org/params/artist"
       |      }
       |	   },
       |    "http://spec.example.org/rels/artistWithOptionalParams": {
@@ -44,14 +45,15 @@ class JsonHomeCacheSpec extends IntegrationSpec {
       |}
     """.stripMargin)
 
-  private val server = JsonHomeHost("http://localhost:8000", Seq(
+  private val port = 8001
+  private val server = JsonHomeHost(s"http://localhost:$port", Seq(
     DirectLinkRelationType("http://spec.example.org/rels/artists"),
     TemplateLinkRelationType("http://spec.example.org/rels/artist"),
     TemplateLinkRelationType("http://spec.example.org/rels/artistWithOptionalParams")
   ))
   private lazy val wsClient = new NingWSClient(new AsyncHttpClientConfig.Builder().build())
   private lazy val actorSystem = ActorSystem(getClass.getSimpleName)
-  private lazy val jsonHomeCache = new JsonHomeCache(server, wsClient, actorSystem)
+  private lazy val jsonHomeCache = AsyncJsonHomeClient.Builder(wsClient).build()
   private var httpServer: HttpServer = _
 
   describe("JsonHomeCache") {
@@ -59,57 +61,22 @@ class JsonHomeCacheSpec extends IntegrationSpec {
     import org.scalatest.concurrent.Eventually._
 
     it("should return Some(link) for href from jsonHomeCache") {
-      jsonHomeCache.getUrl(DirectLinkRelationType("http://spec.example.org/rels/artists")) should be(Some("/artists"))
+      await(jsonHomeCache.getUrl(server, DirectLinkRelationType("http://spec.example.org/rels/artists"))) should be(Some("/artists"))
     }
 
     it("should return None for unknown href relation") {
-      jsonHomeCache.getUrl(DirectLinkRelationType("http://spec.example.org/rels/unknown_relation")) should be(None)
+      await(jsonHomeCache.getUrl(server, DirectLinkRelationType("http://spec.example.org/rels/unknown_relation"))) should be(None)
     }
 
     it("should return Some(link) for href-template from jsonHomeCache") {
-      jsonHomeCache.getUrl(TemplateLinkRelationType("http://spec.example.org/rels/artist")) should be(Some("/artists/{artist-id}"))
-    }
-
-    it("should return None when json home not found from jsonHomeCache") {
-      val server = JsonHomeHost("http://localhost:8001/this_server_and_doc_does_not_exist", Seq())
-      val cache = new JsonHomeCache(server, wsClient, actorSystem)
-      cache.getUrl(DirectLinkRelationType("http://spec.example.org/rels/artists")) should be (None)
-    }
-
-    it("should continuously reload json-home") {
-      val relAlbums = DirectLinkRelationType("http://spec.example.org/rels/albums")
-      val server = JsonHomeHost("http://localhost:8000", Seq(DirectLinkRelationType("http://spec.example.org/rels/artists"), relAlbums))
-      val cache = new JsonHomeCache(server, wsClient, actorSystem, 20 milliseconds)
-
-      eventually {
-        cache.getUrl(relAlbums) should be(None)
-      }
-
-      // let the http server provide a different json-home
-      json = Json.parse(
-        """
-          |{
-          |  "resources": {
-          |    "http://spec.example.org/rels/artists": {
-          |	     "href": "/artists"
-          |	   },
-          |    "http://spec.example.org/rels/albums": {
-          |	     "href": "/albums"
-          |	   }
-          |  }
-          |}
-        """.stripMargin)
-
-      eventually {
-        cache.getUrl(relAlbums) should be(Some("/albums"))
-      }
+      await(jsonHomeCache.getUrl(server, TemplateLinkRelationType("http://spec.example.org/rels/artist"), Map("artist_id" -> 42))) should be(Some("/artists/42"))
     }
 
   }
 
 
   override def beforeAll(configMap: ConfigMap) {
-    httpServer = startServer(8000, JsonHomeHost.jsonHomePath)
+    httpServer = startServer(port, JsonHomeHost.jsonHomePath)
   }
 
   override def afterAll(configMap: ConfigMap) {
